update to infrastructure
This is a partial commit and might not compile on itself due to
signature changes in the methods.

Minor changes here and there to the infrastructure. shared_ptr's have
been dropped in favour of unique_ptr's (not clear it will be better, but
I thought it was the cause of a bug.)

Class member initialisation has been tightened around the place, after
my having spent literally half a day today to hunt down a bug in Octree,
which caused some variables not to be initialised when I expected them
to be (thanks C++!!!!).

Virtual destructors are introduced to base classes as is required for
compilation.

Octree M/L are modified (to M/F) to reflect the fact that multipole
information is actually stored in tensors, not in 1D.

Vectors now find spherical polar coordinats via GetPhi and GetTheta.

New test cases are also added.

Heroes today are valgrind, gdb and -fsanitize=address.