fix integer overflow in FMM
After running the benchmarking, you can imagine my surprise when the FMM
suddenly rose to almost the brute-force time at n = 5e4, after having
been the quickest and largely linear for smaller values of n. Now I
thought to myself, that is a very suspicious number - not unlike the
square root of INT_MAX. But I couldn't figure out what the problem was.

I thought, maybe auto-vectorization has a limit? But surely it works
better for larger values of n? Finally I found my courage to run the
already painfully slow program through valgrind (callgrind). Guess what,
Interact() was only run once!?

That's impossible, there must have been a bug in callgrind. So I defined
counters in FMM to see how many times the different if's were run. Turns
out Interact() really was just run once! And it was the brute-force
statement that was executed!?

And then I remembered the suspicious number thing again, and remembered
we get a (number of particles) ** 2 right at Interact(root, root, grid),
to determine if we should go straight into pairwise interactions if the
product is below a limit. Of course it is, because the product overflew
to a negative number... :)

TL;DR GOOD OLD CLASSIC INTEGER OVERFLOW GOT ME AGAIN AND THAT'S FUNNY.