remove boost dependency, update makefile
Boost was initially used to calculate spherical harmonics. It was too
slow so I replaced it with my own, but kept the functions there so I
could "check" against other methods, or define a fallback if the order
is so high it might be quicker than a recursive method for a single
solid harmonic (would it? Press x to doubt...) Note that the limit was
set to -1 which means we never fallback to boost.

Come to think of it, even at p = 30, a recursive approach (not even
preprocessing) would be as fast as boost for p = 3, so it really makes
no sense to keep that dependency on boost. But now that boost is gone,
I've improved the test file a little bit so I have a little more
confidence in my own methods.

Additionally update makefile.doc to add a delay. Sometimes calling latex
make right after doxygen make seems to cause latex compile issues.